GH-5283: Refactor the folder structure for the service.
