+++ projects/storefrontlib/src/cms-components/asm/asm-main-ui/asm-main-ui.component.spec.ts
@@ -19,7 +19,7 @@ import {
   UserToken,
 } from '@spartacus/core';
 import { Observable, of } from 'rxjs';
-import { AsmComponentService } from '../asm-component.service';
+import { AsmComponentService } from '../services/asm-component.service';
 import { AsmMainUiComponent } from './asm-main-ui.component';
 
 const mockToken = {
