+++ projects/storefrontlib/src/cms-components/asm/asm-session-timer/asm-session-timer.component.spec.ts
@@ -8,7 +8,7 @@ import {
 } from '@angular/core/testing';
 import { AsmConfig, I18nTestingModule, RoutingService } from '@spartacus/core';
 import { of } from 'rxjs';
-import { AsmComponentService } from '../asm-component.service';
+import { AsmComponentService } from '../services/asm-component.service';
 import { AsmSessionTimerComponent } from './asm-session-timer.component';
 import createSpy = jasmine.createSpy;
