+++ projects/storefrontlib/src/cms-components/asm/index.ts
@@ -1,2 +1,9 @@
-export * from './asm.module';
 export * from './asm-main-ui/asm-main-ui.component';
+export * from './asm-root/asm-root.component';
+export * from './asm-session-timer/asm-session-timer.component';
+export * from './asm-session-timer/format-timer.pipe';
+export * from './asm.module';
+export * from './csagent-login-form/csagent-login-form.component';
+export * from './customer-emulation/customer-emulation.component';
+export * from './customer-selection/customer-selection.component';
+export * from './services/index';
