+++ projects/storefrontlib/src/cms-components/asm/services/index.ts
@@ -1,0 +1,1 @@
+export * from './asm-component.service';
